LocalTxs and TxQ honor SeqOrTicket:
The introduction of tickets requires that the rules for the TxQ
change.  The changes affect how sequence-based transactions are
handled by the TxQ as well as incorporating new rules for the
handling of ticket-based transactions.  You can find a long
discourse on the details of the new rules in the TxQ.cpp file.