Return TxConsequences from preflight:
Rather than requesting TxConsequences later in processing,
return TxConsequences from preflight, since the TxConsequences
can be cheaply computed at that time.

Since TxConsequences are cheap to compute, they are no longer
handled with optionals in the TxQ.

This work is being done prior to major surgery on the TxQ so
the two different changes are easier to audit.